# Create an index file while writing a sorted BAM file