This PR is to add a function to create an index file if it is already sorted when creating a BAM file.
PR for issue #141 .