@alumi 2020-06-08
Hi @xckitahara, Thank you for working on this feature!
Could you squash your commits into some concise and meaningful ones and then rebase them onto the master branch?