@alumi 2020-06-08
It seems like GitHub cannot associate your commits with your account @xckitahara.
Please configure your git client with [`git config --global user.email`](https://git-scm.com/book/en/v2/Getting-Started-First-Time-Git-Setup#_your_identity).