@alumi 2020-06-08
> Set up
Great :+1:
Now you need to amend `—reset-author` commits to fix the committer and the author