@xckitahara 2020-06-15
Performance comparison

Time measurement of sort:
① before change,
② after change (with bai created),
③ after change (without bai created)
It was measured at. (Since sort is used for both aln write and block write)
The target file is medium-sam-file used in the test etc.

① time: 761.079912 ms, sd: 98.673396 µs
② time: 925.187657 ms, sd: 47.099024 µs
③ time: 741.354834 ms, sd: 76.992721 µs